s/BatchingAppendSession/Producer/g
